Veritabanı order ile package arasındaki ilişki bire-çok olarak düzenlendi.

+++ MarketOtomasyonu.WFA/Dialogs/GoodsAcceptanceDialogForm.cs
@@ -47,7 +47,7 @@ namespace MarketOtomasyonu.WFA.Dialogs
                         CategoryId = (cmbCategory.SelectedItem as Category).CategoryId,
                          ProductName = txtProductName.Text,
                           ProductBarcode = txtProductBarcode.Text,
-                           ProductPurchasingPrice= Convert.ToInt32(txtProductPurchasingPrice.Text),
+                           ProductPurchasingPrice= Convert.ToDecimal(txtProductPurchasingPrice.Text),
                             ProductSellingPrice=0,
                              ProductStock=0
